+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Gauge component
 - Waterfall component
+- Constructor-schemes: build scheme from search
+- Constructor-schemes: save multiple schemes by token
 
 ### Changed